refactor: #11 Skillsのカスタムフックスを作成しロジックを分離
